+++ src/transform_ocl.cpp
@@ -23,7 +23,7 @@ transform * transform::create_ocl(const uint32_t b, const uint32_t n, const bool
 	// }
 
 	transform * pTransform = nullptr;
-	if (b * static_cast<uint64_t>(b) >= (P1_32 * static_cast<uint64_t>(P2_32) / 2) / (1 << n))
+	if (b * static_cast<uint64_t>(b) >= (P1_32 * static_cast<uint64_t>(P2_32) / 2) / (size_t(1) << n))
 	{
 		pTransform = new transformGPU<3>(b, n, isBoinc, device, num_regs, boinc_platform_id, boinc_device_id, verbose);
 	}
